+++ example/lib/utils/colors.dart
@@ -4,7 +4,4 @@ const Color borderColor = Color(0xffced4da);
 const Color whiteColor = Color(0xffffffff);
 const Color iconColor = Color(0xff000000);
 const Color dividerColor = Color(0xffced4da);
-const Color cursorColor = Color(0xff000000);
-const Color chipsSelectedBackgroundColor = Color(0xff4ab2f1);
-const Color chipsUnselectedBackgroundColor = Color(0xffffffff);
 const Color lightTextColor = Color(0xff212121);