+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## 0.0.1
+## [0.0.1] - 2022-09-05
 
 * TODO: Describe initial release.
+
+## [1.0.2] - 2022-09-07
+### Added
+A number of parameters that allow more customizations.
+- `searchIcon`
+- `unSelectTextStyle`
+- `selectTextStyle`
+- `searchHintTextStyle`, `suggestionListBorderColor`
+- `clearAll`, `confirmText`, `cancelText`